feat: card do perfil funcionando, falta table views
